+++ test/id/sdk.test.ts
@@ -1,10 +1,10 @@
 import axios from "axios";
-import { W3blockIdSDK } from '../src/index';
+import { W3blockIdSDK } from '../../src/index';
 
 import MockAdapter from "axios-mock-adapter";
 
-import { AUTH_REFRESH, AUTH_TOKEN } from "./fixtures/token";
-import { SignInResponseDto } from "../src/w3blockid/api/api";
+import { AUTH_REFRESH, AUTH_TOKEN } from "../fixtures/token";
+import { SignInResponseDto } from "../../src/id/api/api";
 const mock = new MockAdapter(axios);
 
 
@@ -19,6 +19,7 @@ const signInResponse: SignInResponseDto = {
 
 // This sets the mock adapter on the default instance
 mock.onPost("/auth/signin").reply(201, signInResponse);
+mock.onPost("/auth/signin/tenant").reply(201, signInResponse);
 
 describe("SDK", () => {
   const baseURL = process.env.API_BASE_URL || 'http://localhost:6007'
@@ -56,11 +57,12 @@ describe("SDK", () => {
   })
 
   describe("connect app credential", () => {
-    xit("should be able to instantiate", async () => {
+    it("should be able to instantiate", async () => {
       const sdk = new W3blockIdSDK({
         credential: {
           key: 'app-key',
           secret: 'app-secret',
+          tenantId: '5bd96aaf-1888-493e-b84e-7c54a4731868',
         },
         baseURL,
         autoRefresh: false
@@ -69,20 +71,19 @@ describe("SDK", () => {
     }
     );
 
-    xit("should be able to connect", async () => {
+    it("should be able to connect", async () => {
       const sdk = new W3blockIdSDK({
         credential: {
           key: 'app-key',
           secret: 'app-secret',
+          tenantId: '5bd96aaf-1888-493e-b84e-7c54a4731868',
         },
         baseURL,
-        tokens: {
-          authToken: 'auth-token',
-          refreshToken: 'refresh-token',
-        },
         autoRefresh: false
       });
       expect(sdk).toBeDefined();
+      await sdk.connect();
+      expect(sdk.isConnected()).toBeTruthy();
     });
   });
