+++ tools/config.js
@@ -2,7 +2,7 @@ require('dotenv/config');
 
 const path = require('node:path');
 
-const getConfig = () => {
+const getConfig = (context = 'id') => {
   const rootDir = path.resolve('./');
   return {
     path: {
@@ -10,11 +10,11 @@ const getConfig = () => {
     },
     swagger: {
       jsonPath: process.env.SWAGGER_JSON_PATH || 'http://localhost:6007/sdk-json',
-      dest: path.resolve(rootDir, './src/w3blockid/api'),
+      dest: path.resolve(rootDir, `./src/${context}/api`),
       filename: 'swagger.json'
     },
     output: {
-      api: path.resolve(rootDir, './src/w3blockid/api/api.ts'),
+      api: path.resolve(rootDir, `./src/${context}/api/api.ts`),
     }
   }
 };
